pyproject.toml
[project]
name = "invoicegen-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bcrypt>=5.0.0",
    "colorama>=0.4.6",
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.124.4",
    "passlib>=1.7.4",
    "psycopg2>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyotp>=2.9.0",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.45",
    "strawberry-graphql[fastapi]>=0.287.3",
]
